requirements.txt
aiohttp>=3.5.4
websockets>=7.0
pycryptodomex>=3.7
cryptography>=2.6.1
ecdsa>=0.13
base58>=1.0.3
requests>=2.21.0